Extract vix module loading tables
